fix some issues in excutemethod
