+++ src/test/kotlin/interactor/GetTopCitesNameInFshionTaskInteractorTest.kt
@@ -1,5 +1,6 @@
 package interactor
 
+import fakeDataSource.FakeDataSourceForFashionTask
 import model.*
 import org.junit.jupiter.api.*
 import kotlin.test.assertEquals
@@ -7,33 +8,35 @@ import kotlin.test.assertFalse
 import kotlin.test.assertTrue
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-class GetTopFiveCitesNameInClothesFromFamousBrandsInteractorTest {
-    private lateinit var topFiveCitesNameInClothesFromFamousBrandsInteractor: GetTopFiveCitesNameInClothesFromFamousBrandsInteractor
-
+class GetTopCitesNameInFshionTaskInteractorTest {
+    private lateinit var topCitesNameInFashionTaskInteractor: GetTopCitesNameInFashionTaskInteractor
     @BeforeAll
     fun setUp() {
         val fakedata = FakeDataSourceForFashionTask()
-        topFiveCitesNameInClothesFromFamousBrandsInteractor =
-            GetTopFiveCitesNameInClothesFromFamousBrandsInteractor(fakedata)
+        topCitesNameInFashionTaskInteractor =
+            GetTopCitesNameInFashionTaskInteractor(fakedata)
     }
 
     @Test
-    fun shouldTheListSizebe5() {
-//when
-        val result = topFiveCitesNameInClothesFromFamousBrandsInteractor.execute()
-//then
-        assertEquals(5, result.size)
+    fun should_returnFive_when_theListSizeIsFive() {
+        //given
+        //when
+        val result = topCitesNameInFashionTaskInteractor.execute().size
+
+        //then
+        assertEquals(5, result)
     }
 
 
     @Test
-    fun shouldReturnCorrectFiveTopCitiesWheninputisCorrect() {
+    fun should_returnTopFiveCities_when_inputIsCorrect() {
 
         //given
+
         val topCiteNames = listOf("city11", "city12", "city23", "city24", "city47")
 
         //when
-        val actualValue = topFiveCitesNameInClothesFromFamousBrandsInteractor.execute()
+        val actualValue = topCitesNameInFashionTaskInteractor.execute()
 
 
         //then
@@ -42,9 +45,7 @@ class GetTopFiveCitesNameInClothesFromFamousBrandsInteractorTest {
     }
 
     @Test
-    fun shouldReturnCorrectavgWhenifunAvgPriceCorrect() {
-
-
+    fun should_returnCorrectAverage_when_averagePriceIsCorrect() {
         //given
         val clothesPrices = ClothesPrices(
             onePairOfJeansLevis50oneOrSimilar = 170f,
@@ -53,17 +54,17 @@ class GetTopFiveCitesNameInClothesFromFamousBrandsInteractorTest {
             onePairOfMenLeatherBusinessShoes = 200f
         )
         //when
-        val avg = avgPrice(
+        val average = averagePrice(
             clothesPrices
         )
 
         //then
 
-        assertEquals(225f, avg)
+        assertEquals(225f, average)
     }
 
     @Test
-    fun returnFalseWhenCitiesVaveMissingClothesPrices() {
+    fun should_returnFalse_When_ClothesPriceDataIsMissing() {
         //given
 
 
@@ -74,13 +75,13 @@ class GetTopFiveCitesNameInClothesFromFamousBrandsInteractorTest {
             onePairOfMenLeatherBusinessShoes = 200f
         )
          //when
-        val result = topFiveCitesNameInClothesFromFamousBrandsInteractor.excludeNullClothesPrices(clothesPrices)
-        //then
+        val result = excludeNullClothesPrices(clothesPrices)
+         //then
         assertFalse { result }
     }
 
     @Test
-    fun returnNullWhenCitiesVaveMissingClothesPrices() {
+    fun should_returnTrue_When_ClothesPriceDataIsComplete() {
         //given
         val clothesPrices = ClothesPrices(
             onePairOfJeansLevis50oneOrSimilar = 170f,
@@ -89,7 +90,7 @@ class GetTopFiveCitesNameInClothesFromFamousBrandsInteractorTest {
             onePairOfMenLeatherBusinessShoes = 200f
         )
         // when
-        val result = topFiveCitesNameInClothesFromFamousBrandsInteractor.excludeNullClothesPrices(clothesPrices)
+        val result = excludeNullClothesPrices(clothesPrices)
         //then
         assertTrue { result }
     }
